// src/gen/busa_pb.ts
export class Pattern extends Message<Pattern> {
  /**
   * @generated from oneof Pattern.is
   */
  is:
    | {
        /**
         * @generated from field: Pattern.PatternEnum enum = 1;
         */
        value: Pattern_PatternEnum;
        case: 'enum';
      }
    | {
        /**
         * @generated from field: bool bool = 2;
         */
        value: boolean;
        case: 'bool';
      }
    | {
        /**
         * @generated from field: int32 var = 3;
         */
        value: number;
        case: 'var';
      }
    | {
        /**
         * @generated from field: int64 int = 4;
         */
        value: bigint;
        case: 'int';
      }
    | {
        /**
         * @generated from field: string string = 5;
         */
        value: string;
        case: 'string';
      }
    | {
        /**
         * @generated from field: Pattern.Structure structure = 6;
         */
        value: Pattern_Structure;
        case: 'structure';
      }
    | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<Pattern>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'Pattern';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'enum', kind: 'enum', T: proto3.getEnumType(Pattern_PatternEnum), oneof: 'is' },
    { no: 2, name: 'bool', kind: 'scalar', T: 8 /* ScalarType.BOOL */, oneof: 'is' },
    { no: 3, name: 'var', kind: 'scalar', T: 5 /* ScalarType.INT32 */, oneof: 'is' },
    { no: 4, name: 'int', kind: 'scalar', T: 3 /* ScalarType.INT64 */, oneof: 'is' },
    { no: 5, name: 'string', kind: 'scalar', T: 9 /* ScalarType.STRING */, oneof: 'is' },
    { no: 6, name: 'structure', kind: 'message', T: Pattern_Structure, oneof: 'is' },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Pattern {
    return new Pattern().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Pattern {
    return new Pattern().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Pattern {
    return new Pattern().fromJsonString(jsonString, options);
  }

  static equals(
    a: Pattern | PlainMessage<Pattern> | undefined,
    b: Pattern | PlainMessage<Pattern> | undefined,
  ): boolean {
    return proto3.util.equals(Pattern, a, b);
  }
}

/**
 * @generated from enum Pattern.PatternEnum
 */
export enum Pattern_PatternEnum {
  /**
   * @generated from enum value: Unit = 0;
   */
  Unit = 0,

  /**
   * @generated from enum value: Wildcard = 1;
   */
  Wildcard = 1,
}
// Retrieve enum metadata with: proto3.getEnumType(Pattern_PatternEnum)
proto3.util.setEnumType(Pattern_PatternEnum, 'Pattern.PatternEnum', [
  { no: 0, name: 'Unit' },
  { no: 1, name: 'Wildcard' },
]);

/**
 * @generated from message Pattern.Structure
 */
export class Pattern_Structure extends Message<Pattern_Structure> {
  /**
   * @generated from field: string name = 1;
   */
  name = '';

  /**
   * @generated from field: repeated Pattern args = 2;
   */
  args: Pattern[] = [];

  constructor(data?: PartialMessage<Pattern_Structure>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'Pattern.Structure';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'name', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 2, name: 'args', kind: 'message', T: Pattern, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Pattern_Structure {
    return new Pattern_Structure().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Pattern_Structure {
    return new Pattern_Structure().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Pattern_Structure {
    return new Pattern_Structure().fromJsonString(jsonString, options);
  }

  static equals(
    a: Pattern_Structure | PlainMessage<Pattern_Structure> | undefined,
    b: Pattern_Structure | PlainMessage<Pattern_Structure> | undefined,
  ): boolean {
    return proto3.util.equals(Pattern_Structure, a, b);
  }
}

/**
 * @generated from message Rule
 */
export class Rule extends Message<Rule> {
  /**
   * @generated from oneof Rule.is
   */
  is:
    | {
        /**
         * @generated from field: Rule.Index index = 1;
         */
        value: Rule_Index;
        case: 'index';
      }
    | {
        /**
         * @generated from field: Rule.Join join = 2;
         */
        value: Rule_Join;
        case: 'join';
      }
    | {
        /**
         * @generated from field: Rule.Function function = 3;
         */
        value: Rule_Function;
        case: 'function';
      }
    | {
        /**
         * @generated from field: Rule.ChoiceConclusion choice_conclusion = 4;
         */
        value: Rule_ChoiceConclusion;
        case: 'choiceConclusion';
      }
    | {
        /**
         * @generated from field: Rule.DatalogConclusion datalog_conclusion = 5;
         */
        value: Rule_DatalogConclusion;
        case: 'datalogConclusion';
      }
    | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<Rule>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'Rule';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'index', kind: 'message', T: Rule_Index, oneof: 'is' },
    { no: 2, name: 'join', kind: 'message', T: Rule_Join, oneof: 'is' },
    { no: 3, name: 'function', kind: 'message', T: Rule_Function, oneof: 'is' },
    { no: 4, name: 'choice_conclusion', kind: 'message', T: Rule_ChoiceConclusion, oneof: 'is' },
    { no: 5, name: 'datalog_conclusion', kind: 'message', T: Rule_DatalogConclusion, oneof: 'is' },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Rule {
    return new Rule().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Rule {
    return new Rule().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Rule {
    return new Rule().fromJsonString(jsonString, options);
  }

  static equals(
    a: Rule | PlainMessage<Rule> | undefined,
    b: Rule | PlainMessage<Rule> | undefined,
  ): boolean {
    return proto3.util.equals(Rule, a, b);
  }
}
// ...
export class Rule_Join_JoinPattern extends Message<Rule_Join_JoinPattern> {
  /**
   * @generated from field: Rule.Join.JoinPattern.JoinLocation loc = 1;
   */
  loc = Rule_Join_JoinPattern_JoinLocation.Shared;

  /**
   * @generated from field: int32 var = 2;
   */
  var = 0;

  constructor(data?: PartialMessage<Rule_Join_JoinPattern>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'Rule.Join.JoinPattern';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'loc', kind: 'enum', T: proto3.getEnumType(Rule_Join_JoinPattern_JoinLocation) },
    { no: 2, name: 'var', kind: 'scalar', T: 5 /* ScalarType.INT32 */ },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): Rule_Join_JoinPattern {
    return new Rule_Join_JoinPattern().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Rule_Join_JoinPattern {
    return new Rule_Join_JoinPattern().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): Rule_Join_JoinPattern {
    return new Rule_Join_JoinPattern().fromJsonString(jsonString, options);
  }

  static equals(
    a: Rule_Join_JoinPattern | PlainMessage<Rule_Join_JoinPattern> | undefined,
    b: Rule_Join_JoinPattern | PlainMessage<Rule_Join_JoinPattern> | undefined,
  ): boolean {
    return proto3.util.equals(Rule_Join_JoinPattern, a, b);
  }
}

/**
 * @generated from enum Rule.Join.JoinPattern.JoinLocation
 */
export enum Rule_Join_JoinPattern_JoinLocation {
  /**
   * @generated from enum value: Shared = 0;
   */
  Shared = 0,

  /**
   * @generated from enum value: Prefix = 1;
   */
  Prefix = 1,

  /**
   * @generated from enum value: FactArg = 2;
   */
  FactArg = 2,

  /**
   * @generated from enum value: FactValue = 3;
   */
  FactValue = 3,
}
// ...
export enum Rule_Function_Builtin {
  /**
   * @generated from enum value: BOOLEAN_TRUE = 0;
   */
  BOOLEAN_TRUE = 0,

  /**
   * @generated from enum value: BOOLEAN_FALSE = 1;
   */
  BOOLEAN_FALSE = 1,

  /**
   * @generated from enum value: NAT_ZERO = 2;
   */
  NAT_ZERO = 2,

  /**
   * @generated from enum value: NAT_SUCC = 3;
   */
  NAT_SUCC = 3,

  /**
   * @generated from enum value: INT_PLUS = 4;
   */
  INT_PLUS = 4,

  /**
   * @generated from enum value: INT_MINUS = 5;
   */
  INT_MINUS = 5,

  /**
   * @generated from enum value: INT_TIMES = 6;
   */
  INT_TIMES = 6,

  /**
   * @generated from enum value: STRING_CONCAT = 7;
   */
  STRING_CONCAT = 7,

  /**
   * @generated from enum value: EQUAL = 8;
   */
  EQUAL = 8,

  /**
   * @generated from enum value: GT = 9;
   */
  GT = 9,

  /**
   * @generated from enum value: GEQ = 10;
   */
  GEQ = 10,
}
// ...
export class Program extends Message<Program> {
  /**
   * @generated from field: repeated Rule rules = 1;
   */
  rules: Rule[] = [];

  /**
   * @generated from field: repeated string seeds = 2;
   */
  seeds: string[] = [];

  /**
   * @generated from field: repeated string forbids = 3;
   */
  forbids: string[] = [];

  /**
   * @generated from field: repeated string demands = 4;
   */
  demands: string[] = [];

  constructor(data?: PartialMessage<Program>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'Program';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'rules', kind: 'message', T: Rule, repeated: true },
    { no: 2, name: 'seeds', kind: 'scalar', T: 9 /* ScalarType.STRING */, repeated: true },
    { no: 3, name: 'forbids', kind: 'scalar', T: 9 /* ScalarType.STRING */, repeated: true },
    { no: 4, name: 'demands', kind: 'scalar', T: 9 /* ScalarType.STRING */, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Program {
    return new Program().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Program {
    return new Program().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Program {
    return new Program().fromJsonString(jsonString, options);
  }

  static equals(
    a: Program | PlainMessage<Program> | undefined,
    b: Program | PlainMessage<Program> | undefined,
  ): boolean {
    return proto3.util.equals(Program, a, b);
  }
}